Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

Percival.AugLagModel

Percival.percival

Percival.update_cx!

Percival.update_y!

Percival.update_μ!

​

Percival.AugLagModel — Type

AugLagModel(model, y, μ, x, cx)

Given a model

\[\min \ f(x) \quad s.t. \quad c(x) = 0, \quad l ≤ x ≤ u,\]

this new model represents the subproblem of the augmented Lagrangian method

\[\min \ f(x) - yᵀc(x) + \tfrac{1}{2} μ \|c(x)\|^2 \quad s.t. \quad l ≤ x ≤ u,\]

where y is an estimates of the Lagrange multiplier vector and μ is the penalty parameter.

In addition to keeping meta and counters as any NLPModel, an AugLagModel also stores

model: The internal model defining $f$, $c$ and the bounds,

y: The multipliers estimate,

μ: The penalty parameter,

x: Reference to the last point at which the function c(x) was computed,

cx: Reference to c(x),

μc_y: storage for y - μ * cx,

store_Jv and store_JtJv: storage used in hprod!.

Use the functions update_cx!, update_y! and update_μ! to update these values.

Percival.percival — Method

percival(nlp)

Implementation of an augmented Lagrangian method. The following keyword parameters can be passed:

μ: Starting value of the penalty parameter (default: 10.0)

atol: Absolute tolerance used in dual feasibility measure (default: 1e-8)

rtol: Relative tolerance used in dual feasibility measure (default: 1e-8)

ctol: (Absolute) tolerance used in primal feasibility measure (default: 1e-8)

max_iter: Maximum number of iterations (default: 1000)

max_time: Maximum elapsed time in seconds (default: 30.0)

max_eval: Maximum number of objective function evaluations (default: 100000)

subsolver_logger: Logger passed to tron (default: NullLogger)

inity: Initial values of the Lagrangian multipliers

subsolver_kwargs: subsolver keyword arguments as a dictionary

Percival.update_cx! — Method

update_cx!(nlp, x)

Given an AugLagModel, if x != nlp.x, then updates the internal value nlp.cx calling cons on nlp.model. Also updates nlp.μc_y.

Percival.update_y! — Method

update_y!(nlp)

Given an AugLagModel, update nlp.y = -nlp.μc_y and updates nlp.μc_y accordingly.

Percival.update_μ! — Method

update_μ!(nlp, μ)

Given an AugLagModel, updates nlp.μ = μ and nlp.μc_y accordingly.